Add in a 'station_name' tag for every datapoint
This adds a tag for the data in device.name, which refers to what the Vue device is named.

Ideally this tag would have been called 'device_name' to follow the naming patterns in
the data and code, however 'device_name' is already being used for what is called chanName in
code, i.e. the name of the channel on a particular Vue device.  To not break exisitng dashboards
this tag is given a different, non-conflicting name.